app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6

SQLAlchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.10

redis==5.0.8
python-dotenv==1.0.1
PyYAML==6.0.2

pydantic==2.9.2
pydantic_core==2.23.4

numpy==2.2.5
pandas==2.2.3
scikit-learn==1.6.1

requests==2.32.3
python-dateutil==2.9.0.post0

PyMuPDF==1.25.5
pypdf==5.5.0
pdfminer.six==20250506
pypdfium2==4.30.1
pdf2image==1.17.0
pytesseract==0.3.13
opencv-python==4.11.0.86
pillow==11.2.1
python-magic==0.4.27

openai==1.51.2

# Optional / future ML & NLP tooling (kept but can be removed if build too heavy)
transformers==4.51.3
sentence-transformers==4.1.0
torch==2.7.0
torchvision==0.22.0
python-multipart==0.0.9
